掲示板２[tech]

+++ board2/delete.php
@@ -7,11 +7,20 @@ $smarty=new MySmarty();
 
     try{
         $db = getDb();
+        //送られてきたデータを変数に格納
+        $id = $_POST['id'];
+        $user_id = $_POST['user_id'];
 
-        //削除ボタンを押されたpostデータを削除
-        $stt = $db->prepare("DELETE FROM post WHERE id = :id ");
-        $stt->bindValue(':id', $_POST['id'], PDO::PARAM_INT);
-        $stt->execute();
+        if($user_id == $_SESSION['login_id']){
+            //削除ボタンを押されたpostデータを削除
+            $stt = $db->prepare("DELETE FROM post WHERE id = :id ");
+            $stt->bindValue(':id', $id);
+            $stt->execute();
+
+            print "削除しました";
+            $db = NULL;
+
+        }
 
         
     }catch (PDOException $e) {
